Refactor IlUrl
