+++ pillarbox-core-business/src/test/java/ch/srgssr/pillarbox/core/business/integrationlayer/IlUrnTest.kt
@@ -5,9 +5,9 @@
 package ch.srgssr.pillarbox.core.business.integrationlayer
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import ch.srgssr.pillarbox.core.business.integrationlayer.service.ILUrl
-import ch.srgssr.pillarbox.core.business.integrationlayer.service.ILUrl.Companion.toIlUrl
 import ch.srgssr.pillarbox.core.business.integrationlayer.service.IlHost
+import ch.srgssr.pillarbox.core.business.integrationlayer.service.IlUrl
+import ch.srgssr.pillarbox.core.business.integrationlayer.service.IlUrl.Companion.toIlUrl
 import ch.srgssr.pillarbox.core.business.integrationlayer.service.Vector
 import org.junit.runner.RunWith
 import kotlin.test.Test
@@ -17,7 +17,7 @@ import kotlin.test.assertEquals
 class IlUrnTest {
     @Test
     fun testInOut() {
-        val ilUrn = ILUrl(host = IlHost.PROD, urn = "urn:rts:video:12345", vector = Vector.MOBILE)
+        val ilUrn = IlUrl(host = IlHost.PROD, urn = "urn:rts:video:12345", vector = Vector.MOBILE)
         assertEquals(ilUrn, ilUrn.uri.toIlUrl())
     }
 }
